DelegatorsParam: Finish out initial build
This commit along with the previous commits implements
the Delegator Parameter sufficiently to move on.

There are still a couple of TODO's out for this one,
but for the most part it's all ++ and the test cases are
successful.

When this gets to the point where I have implemented the
different CAL-ADDRESS value type, I'll need to come back to this
and figure out how to implement that TODO.

Signed-off-by: Mark Stenglein <mark@stengle.in>