Parameter: Refactoring
I know I told @rhitt I don't use the backspace key, and instead write
perfect code, but this was super awkward.

This commit eliminates an unneeded boolean flag variable as well as some
hacky control structures by throwing the exception when the problem is
detected instead of waiting.

As a bonus, this should help increase speed when dealing with longer
lists of parameter values where one may be invalid.

Signed-off-by: Mark Stenglein <mark@stengle.in>