Parameter: paramValues: Setter Refactor
This commit refactors the paramValues setter to use Array.every and
avoid the performance hit from using forEach.

This commit also adds the isParamValue static validator to the Parameter
class. This is used by the Array.every call, and should probably be
moved eventually to a validator class that should also be built.

Signed-off-by: Mark Stenglein <mark@stengle.in>