DelegateesParam: First pass implementation (#20)
* DelegateesParam: Add bare minimum class and comments

This commit adds the DelegateesParam parameter to the repo. The Param
actually does nothing at the moment but I intend to change that.

Additionally a test class is birthed alongside this parameter that
checks the existence of the file. Important stuff.

Signed-off-by: David Haynes <dhaynes3@gmu.edu>

* DelegateesParam: Complete comments defining DelegateesParam

Essentially a big copy paste from the RFC. It's really helpful that
DelegatorsParam is so similar.

Signed-off-by: David Haynes <dhaynes207@gmail.com>

* DelegateesParam: Complete first pass on implementation

Delegatees function identically to delegators though have some slight
verbage changes. As a result borrow everything from DelegatorParam but
modify where needed.

Implement test spec to verify implementaion.

Closes #2

Signed-off-by: David Haynes <dhaynes207@gmail.com>

* DelegateesParam: whitespace fixes

This commit fixes whitespace on a couple of the files introduced in !20

Signed-off-by: Mark Stenglein <mark@stengle.in>